+++ changelog.md
@@ -5,6 +5,7 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Add support for openSUSE Tumbleweed detection. Closes #133.
 - Do not allow icons on platforms and groups to be empty.
 
 ## [1.6.0 (2024-11-11)](https://github.com/kdeldycke/extra-platforms/compare/v1.5.0...v1.6.0)